Extend ensure_no_std test
