+++ ensure_no_std/src/lib.rs
@@ -1,0 +1,6 @@
+#![no_std]
+
+extern crate alloc;
+
+pub mod no_alloc_imports;
+pub mod with_alloc_imports;
