+++ ensure_no_std/src/with_alloc_imports.rs
@@ -1,18 +1,5 @@
-//! cargo build --target thumbv7em-none-eabihf
-#![no_std]
-#![no_main]
-
-extern crate alloc;
-
-use core::panic::PanicInfo;
-
-use cortex_m_rt::entry;
-use embedded_alloc::LlffHeap as Heap;
-
-#[global_allocator]
-static HEAP: Heap = Heap::empty();
-
-use alloc::{format, vec, vec::Vec};
+use alloc::vec;
+use alloc::vec::Vec;
 use deku::prelude::*;
 
 #[derive(Debug, PartialEq, DekuRead, DekuWrite)]
@@ -28,18 +15,7 @@ struct DekuTest {
     after: Vec<u8>,
 }
 
-#[entry]
-fn main() -> ! {
-    // Initialize the allocator BEFORE you use it
-    {
-        use core::mem::MaybeUninit;
-        const HEAP_SIZE: usize = 1024;
-        static HEAP_MEM: [MaybeUninit<u8>; HEAP_SIZE] = [MaybeUninit::uninit(); HEAP_SIZE];
-        unsafe { HEAP.init(HEAP_MEM.as_ptr() as usize, HEAP_SIZE) }
-    }
-
-    // now the allocator is ready types like Box, Vec can be used.
-
+pub fn rw() {
     #[allow(clippy::unusual_byte_groupings)]
     let test_data: &[u8] = &[0b10101_101, 0x02, 0xBE, 0xEF, 0xff, 0xaa];
     let mut cursor = deku::no_std_io::Cursor::new(test_data);
@@ -59,11 +35,4 @@ fn main() -> ! {
 
     // Test writing
     let _val = val.to_bytes().unwrap();
-
-    loop { /* .. */ }
-}
-
-#[panic_handler]
-fn panic(_: &PanicInfo) -> ! {
-    loop {}
 }
